+++ content/yandex_link_fix_main.js
@@ -1,7 +1,7 @@
 /*
- * ClearURLs (Manifest V3 port)
+ * ScrubURLs (unofficial Manifest V3 port of ClearURLs)
  * Copyright (c) 2017-2025 Kevin Röbert (original ClearURLs)
- * Copyright (c) 2026 ClearURLs MV3 port contributors
+ * Copyright (c) 2026 ScrubURLs contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -36,6 +36,6 @@
             configurable: false
         });
     } catch (e) {
-        console.debug('ClearURLs: Failed to hook _borschik property', e);
+        console.debug('ScrubURLs: Failed to hook _borschik property', e);
     }
 })();